put regexp and literal parser in one file
